task: moves our tools to its own namespace, updates the fsnotify dep
