+++ tools/rebuilder/rebuilder.go
@@ -1,0 +1,4 @@
+// package rebuilder manages hot code reloading for the application.
+// it listen for changes in the application directory and restarts
+// the application when a change is detected.
+package rebuilder
