+++ tools/rebuilder/watcher.go
@@ -1,4 +1,4 @@
-package gloves
+package rebuilder
 
 import (
 	"fmt"
